move to one-off logic
rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED
