+++ controllers/components/dashboard/dashboard_support.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"fmt"
 
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
 	"github.com/opendatahub-io/opendatahub-operator/v2/apis/common"
 	componentApi "github.com/opendatahub-io/opendatahub-operator/v2/apis/components/v1alpha1"
 	dsciv1 "github.com/opendatahub-io/opendatahub-operator/v2/apis/dscinitialization/v1"
@@ -11,8 +13,6 @@ import (
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/cluster"
 	odhtypes "github.com/opendatahub-io/opendatahub-operator/v2/pkg/controller/types"
 	odhdeploy "github.com/opendatahub-io/opendatahub-operator/v2/pkg/deploy"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const (
@@ -100,148 +100,4 @@ func computeComponentName() string {
 
 func GetAdminGroup() string {
 	return adminGroups[cluster.GetRelease().Name]
-}
-
-// TODO: to be removed: https://issues.redhat.com/browse/RHOAIENG-21080
-func updateSpecFields(obj *unstructured.Unstructured, updates map[string][]any) (bool, error) {
-	updated := false
-
-	for field, newData := range updates {
-		existingField, exists, err := unstructured.NestedSlice(obj.Object, "spec", field)
-		if err != nil {
-			return false, fmt.Errorf("failed to get field '%s': %w", field, err)
-		}
-
-		if !exists || len(existingField) == 0 {
-			if err := unstructured.SetNestedSlice(obj.Object, newData, "spec", field); err != nil {
-				return false, fmt.Errorf("failed to set field '%s': %w", field, err)
-			}
-			updated = true
-		}
-	}
-
-	return updated, nil
-}
-
-// TODO: to be removed: https://issues.redhat.com/browse/RHOAIENG-21080
-func getNotebookSizesData() []any {
-	return []any{
-		map[string]any{
-			"name": "Small",
-			"resources": map[string]any{
-				"requests": map[string]any{
-					"cpu":    "1",
-					"memory": "4Gi",
-				},
-				"limits": map[string]any{
-					"cpu":    "3",
-					"memory": "8Gi",
-				},
-			},
-		},
-		map[string]any{
-			"name": "Small",
-			"resources": map[string]any{
-				"requests": map[string]any{
-					"cpu":    "1",
-					"memory": "8Gi",
-				},
-				"limits": map[string]any{
-					"cpu":    "2",
-					"memory": "8Gi",
-				},
-			},
-		},
-		map[string]any{
-			"name": "Medium",
-			"resources": map[string]any{
-				"requests": map[string]any{
-					"cpu":    "3",
-					"memory": "24Gi",
-				},
-				"limits": map[string]any{
-					"cpu":    "6",
-					"memory": "24Gi",
-				},
-			},
-		},
-		map[string]any{
-			"name": "Large",
-			"resources": map[string]any{
-				"requests": map[string]any{
-					"cpu":    "7",
-					"memory": "56Gi",
-				},
-				"limits": map[string]any{
-					"cpu":    "14",
-					"memory": "56Gi",
-				},
-			},
-		},
-		map[string]any{
-			"name": "X Large",
-			"resources": map[string]any{
-				"requests": map[string]any{
-					"cpu":    "15",
-					"memory": "120Gi",
-				},
-				"limits": map[string]any{
-					"cpu":    "30",
-					"memory": "120Gi",
-				},
-			},
-		},
-	}
-}
-
-// TODO: to be removed: https://issues.redhat.com/browse/RHOAIENG-21080
-func getModelServerSizeData() []any {
-	return []any{
-		map[string]any{
-			"name": "Small",
-			"resources": map[string]any{
-				"requests": map[string]any{
-					"cpu":    "1",
-					"memory": "4Gi",
-				},
-				"limits": map[string]any{
-					"cpu":    "2",
-					"memory": "8Gi",
-				},
-			},
-		},
-		map[string]any{
-			"name": "Medium",
-			"resources": map[string]any{
-				"requests": map[string]any{
-					"cpu":    "4",
-					"memory": "8Gi",
-				},
-				"limits": map[string]any{
-					"cpu":    "8",
-					"memory": "10Gi",
-				},
-			},
-		},
-		map[string]any{
-			"name": "Large",
-			"resources": map[string]any{
-				"requests": map[string]any{
-					"cpu":    "6",
-					"memory": "16Gi",
-				},
-				"limits": map[string]any{
-					"cpu":    "10",
-					"memory": "20Gi",
-				},
-			},
-		},
-		map[string]any{
-			"name": "Custom",
-			"resources": map[string]any{
-				"requests": map[string]any{},
-				"limits":   map[string]any{},
-			},
-		},
-	}
 }
